Expand node states from boolean to any struct value
While many problems (and arguably any problem) can be represented as a collection of boolean nodes, some problems are much more naturally represented with more than two states per node.
For example Sudoku is most naturally represented as just a 2D grid of nodes, each of which may be assigned a 1-9 value. Otherwise Sudoku would have to be represented as a 3D array of nodes where each grid cell has 9 nodes behind it and a constraint ensuring that exactly one of those 9 nodes is ever selected at a time. That's a harder problem to think about and computationally more intensive to solve than if we had 1/9th of the nodes and the constraint among them "built in" by virtue of only allowing 1 of the 9 values to be assigned to it.